# tweak too-hot params