See https://github.com/ssb-ngi-pointer/jitdb/issues/118#issuecomment-777096352

This was a clear win, I think.