@arj03 2021-02-11
Sweet! I'm wondering if we should make a benchmark here that uses maxcpu so we can measure this?