@staltz 2021-02-11
Locally I get

**Before**

```
  initial indexing

    ✔ null
    ✔ duration: 886ms

  initial indexing maxcpu 86

    ✔ delete indexes folder to start clean
    ✔ null
    ✔ duration: 5042ms
```

**After**

```
  initial indexing

    ✔ null
    ✔ duration: 930ms

  initial indexing maxcpu 86

    ✔ delete indexes folder to start clean
    ✔ null
    ✔ duration: 4817ms
```

You're right, there's not a big enough difference. Maybe there are two things going on: (1) with 1million records, it's makes a bigger diff, (2) muxrpc weirdness again? (no muxrpc in these benchmarks)