@arj03 2021-02-11
Locally on my machine there is almost no difference. It seems to be almost slower with these changes. I'm getting 24s without this and 25s afterwards. This is creating the author index on a full database (1.5million)