@arj03 2021-02-11
And still as I said in the other comment, I'm learning towards creating these indexes with no max cpu, but then again I don't really use this option.