@staltz 2021-02-11
I think this clearly affects Manyverse desktop (also mobile but I'm running the numbers on desktop), for instance, I deleted `value_author.32prefix` before running the app, I ran it 3 times:

**Before**

```
jitdb missing indexes: [ { type: 'EQUAL', data: { seek: [Function: seekAuthor] ...
jitdb growing index +185ms
itdb growing index +44ms
jitdb growing index +240ms
jitdb growing index +425ms
jitdb growing index +5s
jitdb growing index +12s
jitdb growing index +26s
jitdb log.stream #741 done 1255213 records in 55597ms +12s
```

```
jitdb log.stream #429 done 1255213 records in 64716ms +13s
```

```
jitdb log.stream #893 done 1255213 records in 55267ms +10s
```

**After** setting `wait=90`, `maxPause=300`

```
jitdb log.stream #436 done 1255213 records in 13369ms +2s
```

```
jitdb log.stream #510 done 1255213 records in 9482ms +2s
```

```
jitdb log.stream #979 done 1255213 records in 10026ms +2s
```