@arj03 2021-02-11
Really wierd. I don't really understand why which is probably was is bugging me the most. But very happy that it is such a big improvement :)